Orders data type has been replaced to decimal

+++ LetsDoIt.Moody/src/LetsDoIt.Moody.Web/Entities/Requests/CategoryDetailsUpdateRequest.cs
@@ -4,7 +4,7 @@ namespace LetsDoIt.Moody.Web.Entities.Requests
 {
     public class CategoryDetailsUpdateRequest
     { 
-        public int Order { get; set; }
+        public decimal Order { get; set; }
 
         [Required]
         public byte[] Image { get; set; }
